Move blob image handling from utils/images.ts to utils/blob.ts
Necessary to align with utils structure, also useful for make utils/images.ts more extensible.
