+++ src/types.d.ts
@@ -1,6 +1,8 @@
 import type { i18n } from "i18next";
 import type { CanvasNode } from "obsidian";
 
+export type ImageType = string | TFile;
+
 export interface ElectronWindow extends Window {
     WEBVIEW_SERVER_URL: string;
 }
